pyproject.toml
[project]
name = "lexagent"
version = "0.1.0"
description = "Legal Research AI Agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "openai>=1.35.0",
    "tavily-python>=0.3.3",
    "langfuse>=3.0.0",
    "streamlit>=1.35.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "ruff>=0.4.0",
]
